% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qwt-package.R
\docType{package}
\name{qwt-package}
\alias{qwt}
\alias{qwt-package}
\title{qwt: Quarto Website Template}
\description{
A template for creating websites with Quarto. Includes support for multiple output formats including HTML, DOCX, RevealJS slides, and PDF handouts.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/UCD-SERG/qwt}
  \item Report bugs at \url{https://github.com/UCD-SERG/qwt/issues}
}

}
\author{
\strong{Maintainer}: Your Name \email{your.email@example.com}

}
\keyword{internal}
