% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wai-package.R
\docType{package}
\name{wai-package}
\alias{wai}
\alias{wai-package}
\title{wai: Working with AI}
\description{
UCD-SERG lab notes on working responsibly and effectively with AI coding assistants, published as a Quarto website. Includes support for multiple output formats including HTML, DOCX, RevealJS slides, and PDF handouts.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/d-morrison/wai}
  \item Report bugs at \url{https://github.com/d-morrison/wai/issues}
}

}
\author{
\strong{Maintainer}: UCD-SERG \email{serg@ucdavis.edu}

}
\keyword{internal}
